+++ src/main/java/org/urbcomp/startdb/compress/elf/filedecompressor/ElfOnChimpNFileDecompressor32.java
@@ -22,7 +22,7 @@ public class ElfOnChimpNFileDecompressor32 extends AbstractFileDecompressor{
                 IDecompressor32 decompressor = new ElfOnChimpNDecompressor32(block, 128);
                 List<Float> blockValues = decompressor.decompress();
 
-                for (double element : blockValues) {
+                for (float element : blockValues) {
                     stringBuilder.append(String.valueOf(element)).append(System.lineSeparator());
                 }
